+++ sockeye/checkpoint_decoder.py
@@ -74,6 +74,7 @@ class CheckpointDecoder:
         self.max_output_length_num_stds = max_output_length_num_stds
         self.ensemble_mode = ensemble_mode
         self.beam_size = beam_size
+        self.batch_size = 16
         self.bucket_width_source = bucket_width_source
         self.length_penalty_alpha = length_penalty_alpha
         self.length_penalty_beta = length_penalty_beta
@@ -117,7 +118,7 @@ class CheckpointDecoder:
         models, vocab_source, vocab_target = inference.load_models(self.context,
                                                                    self.max_input_len,
                                                                    self.beam_size,
-                                                                   16,  # batch size
+                                                                   self.batch_size,
                                                                    [self.model],
                                                                    [checkpoint],
                                                                    softmax_temperature=self.softmax_temperature,