+++ CHANGELOG.md
@@ -10,7 +10,7 @@ Note that Sockeye has checks in place to not translate with an old model that wa
 
 Each version section may have have subsections for: _Added_, _Changed_, _Removed_, _Deprecated_, and _Fixed_.
 
-## [1.15.2]
+## [1.15.5]
 ### Added
  - New CLI `sockeye.prepare_data` for preprocessing the training data only once before training,
  potentially splitting large datasets into shards. At training time only one shard is loaded into memory at a time,
@@ -21,6 +21,17 @@ Each version section may have have subsections for: _Added_, _Changed_, _Removed
  ```--prepared-data``` argument pointing to the folder containing the preprocessed and sharded data. Using the raw
  training data is still possible and now consumes less memory.
 
+## [1.15.4]
+### Added
+ - A warning will be printed if the checkpoint decoder slows down training.
+
+## [1.15.3]
+### Added
+ - Exposing the xavier random number generator through --weight-init-xavier-rand-type.
+
+## [1.15.2]
+### Added
+ - Exposing MXNet's Nesterov Accelerated Gradient, Adadelta and Adadelta optimizers.
 
 ## [1.15.1]
 ### Added
@@ -115,6 +126,9 @@ sockeye.evaluate now accepts `bleu` and `chrf` as values for `--metrics`
 ### Added
  - Lexicon-based target vocabulary restriction for faster decoding. New CLI for top-k lexicon creation, sockeye.lexicon.
  New translate CLI argument `--restrict-lexicon`.
+
+### Changed
+ - Bleu computation based on Sacrebleu.
 
 ## [1.10.5]
 ### Fixed